% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets_list}
\alias{datasets_list}
\title{List files on Faculty datasets.}
\usage{
datasets_list(prefix = "/", project_id = NULL, show_hidden = FALSE)
}
\arguments{
\item{prefix}{The prefix by which to filter files (default: '/')}

\item{project_id}{The ID of the project; by default, the current project is used.}

\item{show_hidden}{Whether to show hidden files (files prefixed with a dot).}
}
\description{
List files on Faculty datasets.
}
