% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets_put}
\alias{datasets_put}
\title{Copy from the local filesystem to SherlockML datasets.}
\usage{
datasets_put(local_path, datasets_path, project_id = NULL)
}
\arguments{
\item{local_path}{The destination path on the local filesystem.}

\item{datasets_path}{File path on SherlockML datasets}

\item{project_id}{The ID of the project; by default, the current project is used.}
}
\description{
Copy from the local filesystem to SherlockML datasets.
}
