% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report}
\alias{report}
\title{Rmarkdown format for a SherlockML report}
\usage{
report(...)
}
\description{
Knits a file to HTML and makes that HTML available as a SherlockML report
}
\details{
Simply add \code{output: rsherlockml::report} to your Rmd yaml.

\code{rmarkdown::html_document}
}
