% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets_copy}
\alias{datasets_copy}
\title{Copy a file from one location to another on SherlockML Datasets.}
\usage{
datasets_copy(source_path, destination_path, project_id = NULL)
}
\arguments{
\item{source_path}{Source path on SherlockML datasets}

\item{destination_path}{Destination path on SherlockML datasets}

\item{project_id}{The ID of the project; by default, the current project is used.}
}
\description{
Copy a file from one location to another on SherlockML Datasets.
}
