% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report}
\alias{report}
\title{Rmarkdown format for a SherlockML report}
\usage{
report(..., quiet = TRUE, theme = NULL, mathjax = NULL,
  highlight = NULL)
}
\arguments{
\item{...}{all parameters that can be passed to \code{rmarkdown::html_document}}
}
\description{
Knits a file to HTML and makes that HTML available as a Sherlock report
}
\details{
Simply add \code{output: rsherlock::report} to your Rmd yaml.
}
